Rename : useDraftStore를 shared에서 post 내부 store로 이동 지역화
