+++ views/post/post-draft/presentational/PostDraftListPres.tsx
@@ -1,7 +1,7 @@
 import { TrashIcon, InfoCircledIcon } from '@radix-ui/react-icons';
 
 import styles from '../styles/PostDraftListPres.module.scss';
-import { useDraftStore } from '@/views/post-draft/stores/useDraftStore';
+import { useDraftStore } from '@/views/post/post-draft/stores/useDraftStore';
 
 type Props = {
   onDelete: (id: number) => void;
